+++ src/Config.php
@@ -7,4 +7,6 @@ namespace Craftzing\Laravel\LokaliseWebhooks;
 interface Config
 {
     public function lokaliseXSecret(): string;
+
+    public function areIpRestrictionsEnabled(): bool;
 }