+++ src/IlluminateConfig.php
@@ -10,23 +10,30 @@ use Illuminate\Contracts\Config\Repository;
 final class IlluminateConfig implements Config
 {
     private string $lokaliseXSecret;
+    private bool $areIpRestrictionsEnabled;
 
     public function __construct(Repository $config)
     {
-        $this->lokaliseXSecret = $config->get('lokalise-webhooks.x_secret') ?: '';
-
-        $this->guardAgainstMissingLokaliseXSecret();
+        $this->lokaliseXSecret = $this->lokaliseSecretFromConfig($config);
+        $this->areIpRestrictionsEnabled = $config->get('lokalise-webhooks.enable_ip_restrictions');
     }
 
-    private function guardAgainstMissingLokaliseXSecret(): void
+    private function lokaliseSecretFromConfig(Repository $config): string
     {
-        if (! $this->lokaliseXSecret) {
-            throw AppMisconfigured::missingLokaliseXSecret();
+        if ($value = $config->get('lokalise-webhooks.x_secret')) {
+            return $value;
         }
+
+        throw AppMisconfigured::missingLokaliseXSecret();
     }
 
     public function lokaliseXSecret(): string
     {
         return $this->lokaliseXSecret;
+    }
+
+    public function areIpRestrictionsEnabled(): bool
+    {
+        return $this->areIpRestrictionsEnabled;
     }
 }